+++ lib/fishstick_dart.dart
@@ -3,12 +3,10 @@ import "client/client.dart";
 export "commands/info/ping.dart";
 export "commands/info/invite.dart";
 export "commands/info/info.dart";
-
 export "commands/autopost_stw_alerts/autopost.dart";
 export "commands/general/color.dart";
-
 export "commands/premium/premium.dart";
-
 export "commands/partner/partner.dart";
+export "commands/blacklist/blacklist.dart";
 
 final Client client = Client();